fix: enforce correct owner of socket

+++ main.go
@@ -177,9 +177,23 @@ func startServer() {
 }
 
 func checkServer() {
-	if _, err := os.Stat(gSocketPath); os.IsNotExist(err) {
+	st, err := os.Stat(gSocketPath)
+	if os.IsNotExist(err) {
 		startServer()
-	} else if _, err := net.Dial("unix", gSocketPath); err != nil {
+		return
+	}
+	if err != nil {
+		log.Printf("stat socket: %s", err)
+		return
+	}
+
+	// Fail closed if the socket isn't owned by process
+	if !socketOwnedByCurrentUser(st) {
+		log.Printf("refusing to use socket not owned by current user: %s", gSocketPath)
+		return
+	}
+
+	if _, err := net.Dial("unix", gSocketPath); err != nil {
 		if err := os.Remove(gSocketPath); err != nil {
 			log.Print(err)
 		}
